requirements.txt
google-generativeai>=0.8.0
exa-py>=1.0.0
asyncio
aiohttp>=3.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0
typing-extensions>=4.8.0